Fix indendation of parameter list in module declaration
This requires fixing verilog-beg-of-statement-1 that can't deal (find)
the module beginning when there is a parameter list. For that we tell it
to keep on searching when in a parameter list until it finds something
else.

Furthermore when verilog-do-indent encounters #( within the search
boundary it shouldn't align to the current column.